Create Responding.app's AppSupport directory if not there
In order to save the txt files that contain the messages and
responses, it is important to create the directory in which
these files will go. In this case, we create:
  ~/Library/Application Support/Responding

It is in this directory that we store "message_from*.txt" and
"message_to*.txt" and must exist for state-saving to work.